requirements.txt
flask
flask-cors
flask-pymongo
flask-bcrypt
flask-jwt-extended
flask-mail
pandas
numpy
scikit-learn
joblib
shap
xgboost
google-generativeai
python-dotenv
waitress
pyotp
twilio
cloudinary
